ModalViewControllerに処理を移動する
